% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{accuracy_by_class}
\alias{accuracy_by_class}
\title{Title}
\usage{
accuracy_by_class(mod, dat_train, dat_test)
}
\arguments{
\item{mod}{}

\item{dat_train}{}

\item{dat_test}{}
}
\value{

}
\description{
Title
}
