% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_alcohol}
\alias{data_alcohol}
\title{From Urban,Gates 2021 https://doi.org/10.1037/met0000374}
\usage{
data_alcohol(n, extended = TRUE)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
From Urban,Gates 2021 https://doi.org/10.1037/met0000374
}
