% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{evaluate_mod}
\alias{evaluate_mod}
\title{Title}
\usage{
evaluate_mod(mod, dat_train, dat_test, measures)
}
\arguments{
\item{mod}{}

\item{dat_train}{}

\item{dat_test}{}
}
\value{

}
\description{
Title
}
