% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{train_nn}
\alias{train_nn}
\title{Title}
\usage{
train_nn(
  mod,
  x,
  y,
  loss,
  epochs = 20,
  learning_rate = 0.001,
  batch_size = nrow(x),
  metrics = NULL
)
}
\arguments{
\item{mod}{Keras Model, z.B. erstellt mit keras_model_sequential()}

\item{x}{Prädiktoren}

\item{y}{Kriterium}

\item{loss}{Loss Function, Angabe als character, z.B. "mse"}

\item{epochs}{Anzahl der Schritte im Gradient Descent. Default ist 20.}

\item{learning_rate}{Schrittweite im Gradient Descent. Default ist .001.}

\item{batch_size}{Für Stochastic Gradient Descent: Anzahl der Samples in einem Batch}

\item{metrics}{(Optional) Weitere Metriken die neben der Loss beim Training angezeigt werden sollen. Diese werden beim Training nicht berücksichtigt.}
}
\value{

}
\description{
Title
}
